Refactor: Split schema.test.ts (757 lines) into per-schema-section test files alongside schema.ts and extract shared config fixtures into schema.test-helpers.ts (warren-3730)
