+++ .mulch/expertise/warren-config.jsonl
@@ -35,3 +35,4 @@
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-23T00:29:15.162Z","evidence":{"seeds":"warren-cd37","commit":"9af562d480d58d842b5576dba7477d738f389954"},"dir_anchors":["src/warren-config"],"content":"src/warren-config/schema.ts DefaultsConfigSchema.agent.pauseTimeoutMs (warren-cd37 / pl-0344 step 2): nested optional agent block carrying pauseTimeoutMs (z.number().int().min(1_000).max(86_400_000).default(1_800_000)). Bounds 1s..24h, default 30min. Constant DEFAULT_AGENT_PAUSE_TIMEOUT_MS exported alongside for the missing-agent-block fallback. Schema applies .default() so a parsed agent block always carries a number; consumers reading from loadWarrenConfig() without an agent block fall back to the constant. Nested under agent.* (vs flat top-level) so future agent-runtime knobs can grow there without inflating DefaultsConfigSchema's top level. Re-exported from src/warren-config/index.ts. Documented in SPEC §11.O interactive-runs addendum and warren CLAUDE.md per-project config section.","id":"mx-37127e"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-23T15:40:29.663Z","evidence":{"seeds":"warren-b802","commit":"3d56159d48f3c9d9152edb431bdd2f05f2dcf262"},"content":"DefaultsConfigSchema.interactiveAgents (warren-b802) is an optional strict block with brainstormRuntime / plannerRuntime fields validated against KNOWN_RUNTIME_IDS ('claude-code' | 'sapling' | 'pi'). interactiveRuntimeOverride(agentName, defaults) maps agent name → config field; readRuntimeId(agent, configOverride?) resolves at dispatch time with precedence config > frontmatter.runtime > agent.name. The agent row stays 'builtin' — no seed-time mutation.","id":"mx-843b14"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-23T16:45:18.358Z","evidence":{"commit":"d8164d7054ab28c2bffbfdeb8a5e2c2983ed614e"},"dir_anchors":["src/warren-config"],"content":"DefaultsConfigSchema.plotSync (warren-cd22) is an optional strict block with mergeStrategy ('immediate'|'auto'|'manual') and targetBranch to configure plot sync.","id":"mx-2a0894"}
+{"type":"pattern","classification":"tactical","recorded_at":"2026-05-28T04:19:14.736Z","evidence":{"commit":"191025c3a76159f156034cf72b8e1693c687f869"},"name":"splitting-schema-tests","description":"Split the schema.test.ts (757 lines) into schema.triggers.test.ts, schema.defaults.test.ts, and schema.preview.test.ts. Extracted shared config fixtures to schema.test-helpers.ts. Removed the schema.test.ts entry from scripts/file-size-budgets.json.","files":[".seeds/issues.jsonl","scripts/file-size-budgets.json","src/warren-config/schema.test.ts"],"id":"mx-a16edc"}
