+++ scripts/file-size-budgets.json
@@ -31,7 +31,6 @@
 		"src/server/handlers/plots.ts": 1010,
 		"src/server/handlers/plots.workbench.test.ts": 766,
 		"src/server/integration.multi-worker.test.ts": 553,
-		"src/server/server.test.ts": 536,
-		"src/warren-config/schema.test.ts": 757
+		"src/server/server.test.ts": 536
 	}
 }
